src/requirements.txt
langgraph 
transformers
torch 
python-dotenv
fastapi
uvicorn